face-recognition/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
deprecation==2.1.0
dnspython==2.7.0
easydict==1.13
fastapi==0.115.13
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
gotrue==2.12.2
grpcio==1.67.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
insightface==0.7.3
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.3
milvus-lite==2.4.12
mpmath==1.3.0
multidict==6.5.0
networkx==3.5
numpy==2.3.1
onnx==1.18.0
onnxruntime==1.22.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pluggy==1.6.0
postgrest==1.1.1
prettytable==3.16.0
propcache==0.3.2
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymilvus==2.5.11
pymongo==4.13.2
pyparsing==3.2.3
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.5.1
requests==2.32.4
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
simsimd==6.4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
storage3==0.12.0
StrEnum==0.4.15
stringzilla==3.12.5
supabase==2.16.0
supafunc==0.10.1
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.3
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
